requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
chardet==5.2.0
click==8.1.8
deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
flasgger==0.9.7.1
flask==3.1.0
flask-cors==5.0.1
flask-jwt-extended==4.7.1
flask-limiter==3.12
flask-migrate==4.1.0
flask-sqlalchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
limits==4.6
mako==1.3.9
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mistune==3.1.2
ordered-set==4.1.0
packaging==24.2
passlib==1.7.4
pillow==11.1.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
pyjwt==2.10.1
pytest==8.3.5
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pyyaml==6.0.2
redis==5.2.1
referencing==0.36.2
reportlab==4.3.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.39
starlette==0.46.1
typing-extensions==4.12.2
uvicorn==0.34.0
werkzeug==3.1.3
wrapt==1.17.2
flask-mail
